Added fluid handler & moved everything to there.
Still getting error though (when I comment out the addCellHandler function in Fluidcraft.java no error)
